third_party/blink/web_tests/http/tests/origin_trials/webexposed/webcomponentsv0-basic-import-enabled.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate this token with the command:
generate_token.py http://127.0.0.1:8000 WebComponentsV0 --expire-timestamp=2000000000
-->
<meta http-equiv="origin-trial" content="Ar9dxxaQ+S9WJDVAL5PqZEtGNqgN2jeewmtzi1+9TK2oavYps7dKu36815FgbztKvcJJ2QATr2g0k0U6nQaDOAMAAABXeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2ViQ29tcG9uZW50c1YwIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9" />

<link id="target" rel="import" href="resources/html-imports-hello.html">

<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>
<script>
  test(() => {
  	  assert_equals(target.import.querySelector('h1').textContent, 'Hello');
    }, 'Basic html file should be imported');
</script>
